Fix and improve the metadata parser
The parser was broken, only storing the last line of the metadata block.
Now, the metadata block is located first, then parsed.  The parser
recognises the @source and @type lines, specifying the source of a
scroll and its type, respectively.  All the other lines are treated as
comma separated lists of tags.  The case of the tags is ignored, so they
are all converted to lower case.