Switch to a new back-end based on bleve
The back-end is very basic, treating the scrolls as a single chunk of
data, just like the old swish++ back-end.  In contrast to swish++, bleve
allows searching for Unicode strings.  For example, swish++ did not find
any results for "Kähler".

In the future, the bleve back-end could provide search based on metadata
like searching for scrolls of type 'lemma' from a source matching
'huybrechts' that have not been tagged as 'decomposition' and which
contain a certain word.