Properly handle errors during rendering
Previously, XeLaTeX would silently fail, now an error is created if a
scroll cannot be rendered. Also, there is now an early return in all
methods of xelatexImagemagickRenderer, so x.error is not changed after
an error occurred.

Note that the errors were not passed along properly, because the methods
were called on *different copies* of the renderer, because they were
implemented on a value, not a reference.